release: v0.3.1 and fix broken tests in go 1.13
Releases a new version `v0.3.1` and fixes the broken tests in go 1.13,
see https://github.com/golang/go/issues/33774.

Additionally, does some minor tidy of the installed Go binaries.
