+++ go.mod
@@ -1,1 +1,9 @@
 module github.com/elastic/go-licenser
+
+go 1.13
+
+require (
+	golang.org/x/lint v0.0.0-20200302205851-738671d3881b // indirect
+	golang.org/x/mod v0.3.0 // indirect
+	golang.org/x/tools v0.0.0-20200601175630-2caf76543d99 // indirect
+)
